# attached_assets/wheatrisk_build/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WheatRisk.ai API - District-level wheat crop climate risk data for Uttar Pradesh
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RiskDna {
  drought: number;
  tail_risk: number;
  volatility: number;
  floor: number;
  premium: number;
}

export type RiskMetricsLevel = typeof RiskMetricsLevel[keyof typeof RiskMetricsLevel];


export const RiskMetricsLevel = {
  Very_Low: 'Very Low',
  Low: 'Low',
  Moderate: 'Moderate',
  High: 'High',
  Very_High: 'Very High',
} as const;

export interface RiskMetrics {
  level: RiskMetricsLevel;
  mean_chi: number;
  p95_tail: number;
  p5_floor: number;
  volatility: number;
  dna: RiskDna;
}

export type RainfallDataCategory = typeof RainfallDataCategory[keyof typeof RainfallDataCategory];


export const RainfallDataCategory = {
  Excess: 'Excess',
  Normal: 'Normal',
  Deficient: 'Deficient',
  Scanty: 'Scanty',
  No_Rain: 'No Rain',
} as const;

export interface RainfallData {
  actual_mm: number;
  normal_mm: number;
  deviation_pct: number;
  category: RainfallDataCategory;
}

export type TemperatureDataCategory = typeof TemperatureDataCategory[keyof typeof TemperatureDataCategory];


export const TemperatureDataCategory = {
  Cooler: 'Cooler',
  Normal: 'Normal',
  Warm: 'Warm',
  Hot: 'Hot',
  Extreme: 'Extreme',
} as const;

export interface TemperatureData {
  actual_c: number;
  normal_c: number;
  anomaly_c: number;
  category: TemperatureDataCategory;
}

export interface TrajectoryPoint {
  stage: string;
  actual: number;
  normal: number;
}

export interface Recommendation {
  text: string;
  pd_adjustment_pct: number;
  lgd_adjustment_pct: number;
}

export interface DistrictRisk {
  district: string;
  state: string;
  season: string;
  growth_stage: string;
  risk: RiskMetrics;
  rainfall: RainfallData;
  temperature: TemperatureData;
  trajectory: TrajectoryPoint[];
  recommendation: Recommendation;
}

export interface DistributionEntry {
  category: string;
  count: number;
  percentage: number;
}

export interface DistributionStats {
  total_districts: number;
  favorable_count: number;
  elevated_count: number;
  distribution: DistributionEntry[];
}

export interface GrowthStage {
  id: string;
  label: string;
  period: string;
}

export type ListDistrictsParams = {
season?: string;
growth_stage?: string;
layer?: ListDistrictsLayer;
};

export type ListDistrictsLayer = typeof ListDistrictsLayer[keyof typeof ListDistrictsLayer];


export const ListDistrictsLayer = {
  wheat_risk: 'wheat_risk',
  drought_stress: 'drought_stress',
  heat_stress: 'heat_stress',
} as const;

export type FetchDistrictDetailParams = {
district: string;
season?: string;
growth_stage?: string;
};

export type GetDistributionParams = {
season?: string;
growth_stage?: string;
layer?: GetDistributionLayer;
};

export type GetDistributionLayer = typeof GetDistributionLayer[keyof typeof GetDistributionLayer];


export const GetDistributionLayer = {
  wheat_risk: 'wheat_risk',
  drought_stress: 'drought_stress',
  heat_stress: 'heat_stress',
} as const;