# attached_assets/wheatrisk_build/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WheatRisk.ai API - District-level wheat crop climate risk data for Uttar Pradesh
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: districts
    description: District risk data
  - name: filters
    description: Filter options
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /districts:
    get:
      operationId: listDistricts
      tags: [districts]
      summary: List all districts with risk data
      parameters:
        - name: season
          in: query
          schema:
            type: string
        - name: growth_stage
          in: query
          schema:
            type: string
        - name: layer
          in: query
          schema:
            type: string
            enum: [wheat_risk, drought_stress, heat_stress]
      responses:
        "200":
          description: List of districts with risk data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DistrictRisk"

  /districts/detail:
    get:
      operationId: fetchDistrictDetail
      tags: [districts]
      summary: Get detailed data for a single district
      parameters:
        - name: district
          in: query
          required: true
          schema:
            type: string
        - name: season
          in: query
          schema:
            type: string
        - name: growth_stage
          in: query
          schema:
            type: string
      responses:
        "200":
          description: District risk detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistrictRisk"
        "404":
          description: District not found

  /districts/distribution:
    get:
      operationId: getDistribution
      tags: [districts]
      summary: Get distribution stats across all districts for the active layer
      parameters:
        - name: season
          in: query
          schema:
            type: string
        - name: growth_stage
          in: query
          schema:
            type: string
        - name: layer
          in: query
          schema:
            type: string
            enum: [wheat_risk, drought_stress, heat_stress]
      responses:
        "200":
          description: Distribution statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistributionStats"

  /filters/seasons:
    get:
      operationId: listSeasons
      tags: [filters]
      summary: Get available seasons
      responses:
        "200":
          description: Available seasons
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /filters/growth-stages:
    get:
      operationId: listGrowthStages
      tags: [filters]
      summary: Get available wheat growth stages
      responses:
        "200":
          description: Available growth stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrowthStage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RiskDna:
      type: object
      properties:
        drought:
          type: number
        tail_risk:
          type: number
        volatility:
          type: number
        floor:
          type: number
        premium:
          type: number
      required: [drought, tail_risk, volatility, floor, premium]

    RiskMetrics:
      type: object
      properties:
        level:
          type: string
          enum: [Very Low, Low, Moderate, High, Very High]
        mean_chi:
          type: number
        p95_tail:
          type: number
        p5_floor:
          type: number
        volatility:
          type: number
        dna:
          $ref: "#/components/schemas/RiskDna"
      required: [level, mean_chi, p95_tail, p5_floor, volatility, dna]

    RainfallData:
      type: object
      properties:
        actual_mm:
          type: number
        normal_mm:
          type: number
        deviation_pct:
          type: number
        category:
          type: string
          enum: [Excess, Normal, Deficient, Scanty, No Rain]
      required: [actual_mm, normal_mm, deviation_pct, category]

    TemperatureData:
      type: object
      properties:
        actual_c:
          type: number
        normal_c:
          type: number
        anomaly_c:
          type: number
        category:
          type: string
          enum: [Cooler, Normal, Warm, Hot, Extreme]
      required: [actual_c, normal_c, anomaly_c, category]

    TrajectoryPoint:
      type: object
      properties:
        stage:
          type: string
        actual:
          type: number
        normal:
          type: number
      required: [stage, actual, normal]

    Recommendation:
      type: object
      properties:
        text:
          type: string
        pd_adjustment_pct:
          type: number
        lgd_adjustment_pct:
          type: number
      required: [text, pd_adjustment_pct, lgd_adjustment_pct]

    DistrictRisk:
      type: object
      properties:
        district:
          type: string
        state:
          type: string
        season:
          type: string
        growth_stage:
          type: string
        risk:
          $ref: "#/components/schemas/RiskMetrics"
        rainfall:
          $ref: "#/components/schemas/RainfallData"
        temperature:
          $ref: "#/components/schemas/TemperatureData"
        trajectory:
          type: array
          items:
            $ref: "#/components/schemas/TrajectoryPoint"
        recommendation:
          $ref: "#/components/schemas/Recommendation"
      required: [district, state, season, growth_stage, risk, rainfall, temperature, trajectory, recommendation]

    DistributionEntry:
      type: object
      properties:
        category:
          type: string
        count:
          type: number
        percentage:
          type: number
      required: [category, count, percentage]

    DistributionStats:
      type: object
      properties:
        total_districts:
          type: number
        favorable_count:
          type: number
        elevated_count:
          type: number
        distribution:
          type: array
          items:
            $ref: "#/components/schemas/DistributionEntry"
      required: [total_districts, favorable_count, elevated_count, distribution]

    GrowthStage:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        period:
          type: string
      required: [id, label, period]