# attached_assets/wheatrisk_build/lib/api-zod/src/generated/types/distributionStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WheatRisk.ai API - District-level wheat crop climate risk data for Uttar Pradesh
 * OpenAPI spec version: 0.1.0
 */
import type { DistributionEntry } from './distributionEntry';

export interface DistributionStats {
  total_districts: number;
  favorable_count: number;
  elevated_count: number;
  distribution: DistributionEntry[];
}