# attached_assets/wheatrisk_build/lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WheatRisk.ai API - District-level wheat crop climate risk data for Uttar Pradesh
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}