# lib/api-zod/src/generated/types/distributionEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WheatRisk.ai API - District-level wheat crop climate risk data for Uttar Pradesh
 * OpenAPI spec version: 0.1.0
 */

export interface DistributionEntry {
  category: string;
  count: number;
  percentage: number;
}