# lib/api-zod/src/generated/types/rainfallData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WheatRisk.ai API - District-level wheat crop climate risk data for Uttar Pradesh
 * OpenAPI spec version: 0.1.0
 */
import type { RainfallDataCategory } from './rainfallDataCategory';

export interface RainfallData {
  actual_mm: number;
  normal_mm: number;
  deviation_pct: number;
  category: RainfallDataCategory;
}